+++ app/schemas/enums.py
@@ -1,4 +1,4 @@
-from enum import Enum
+from enum import Enum, IntEnum
 
 from .gameenums import (
     AI_ACT_NUM_NAME,
@@ -11,12 +11,17 @@ from .gameenums import (
     CARD_TYPE_NAME,
     CLASS_OVERWRITE_NAME,
     COND_TYPE_NAME,
+    DETAIL_MISSION_COND_TYPE,
+    DETAIL_MISSION_LINK_TYPE,
     EVENT_TYPE_NAME,
     FUNC_TARGETTYPE_NAME,
     FUNC_TYPE_NAME,
     GENDER_TYPE_NAME,
     GIFT_TYPE_NAME,
     ITEM_TYPE_NAME,
+    MISSION_PROGRESS_TYPE_NAME,
+    MISSION_REWARD_TYPE_NAME,
+    MISSION_TYPE_NAME,
     PAY_TYPE_NAME,
     PURCHASE_TYPE_NAME,
     QUEST_CONSUME_TYPE_NAME,
@@ -42,12 +47,17 @@ from .gameenums import (
     NiceClassRelationOverwriteType,
     NiceCondType,
     NiceConsumeType,
+    NiceDetailMissionCondLinkType,
+    NiceDetailMissionCondType,
     NiceEventType,
     NiceFuncTargetType,
     NiceFuncType,
     NiceGender,
     NiceGiftType,
     NiceItemType,
+    NiceMissionProgressType,
+    NiceMissionRewardType,
+    NiceMissionType,
     NicePayType,
     NicePurchaseType,
     NiceQuestType,
@@ -404,6 +414,36 @@ AI_TIMING_NAME: dict[int, AiTiming] = {
     5: AiTiming.turnEnemyStart,
     6: AiTiming.turnEnemyEnd,
 }
+
+
+### Mission Cond Detail Type ###
+
+
+class DetailMissionCondType(IntEnum):
+    ENEMY_KILL_NUM = 1
+    ENEMY_INDIVIDUALITY_KILL_NUM = 2
+    ITEM_GET_TOTAL = 3
+    BATTLE_SVT_IN_DECK = 4  # Unused
+    BATTLE_SVT_EQUIP_IN_DECK = 5  # Unused
+    TARGET_QUEST_ENEMY_KILL_NUM = 6
+    TARGET_QUEST_ENEMY_INDIVIDUALITY_KILL_NUM = 7
+    TARGET_QUEST_ITEM_GET_TOTAL = 8
+    QUEST_CLEAR_ONCE = 9
+    QUEST_CLEAR_NUM_1 = 10
+    ITEM_GET_BATTLE = 12
+    DEFEAT_ENEMY_INDIVIDUALITY = 13
+    DEFEAT_ENEMY_CLASS = 14
+    DEFEAT_SERVANT_CLASS = 15
+    DEFEAT_ENEMY_NOT_SERVANT_CLASS = 16
+    BATTLE_SVT_CLASS_IN_DECK = 18  # Filter by svt class
+    SVT_GET_BATTLE = 19  # Embers are svt instead of items
+    FRIEND_POINT_SUMMON = 21
+    BATTLE_SVT_ID_IN_DECK = 23  # Filter by svt ID
+    QUEST_CLEAR_NUM_2 = 24  # Not sure what's the difference QUEST_CLEAR_NUM_1
+    DICE_USE = 25  # Probably Fate/Requiem event
+    SQUARE_ADVANCED = 26
+    MORE_FRIEND_FOLLOWER = 27  # 5th Anniversary missions
+    MAIN_QUEST_DONE = 28  # 22M Download Campaign
 
 
 ### Trait ###
@@ -1030,4 +1070,14 @@ __all__ = [
     "AiTiming",
     "AI_TIMING_NAME",
     "AiType",
+    "NiceMissionType",
+    "MISSION_TYPE_NAME",
+    "NiceMissionRewardType",
+    "MISSION_REWARD_TYPE_NAME",
+    "NiceMissionProgressType",
+    "MISSION_PROGRESS_TYPE_NAME",
+    "NiceDetailMissionCondType",
+    "DETAIL_MISSION_COND_TYPE",
+    "NiceDetailMissionCondLinkType",
+    "DETAIL_MISSION_LINK_TYPE",
 ]